refactored employee dash component, it was getting too big
